requirements.txt
# Основные зависимости для поиска экзопланет
numpy>=1.20.0
scipy>=1.7.0
pandas>=1.3.0
matplotlib>=3.3.0
seaborn>=0.11.0

# Машинное обучение
torch>=2.0.0
scikit-learn>=0.24.0

# Астрономические библиотеки
astropy>=5.0.0
lightkurve>=2.0.0
astroquery>=0.4.6

# Утилиты
tqdm>=4.50.0
PyYAML>=6.0.0
psutil>=5.8.0

# Дополнительные библиотеки (опционально)
# batman-package>=2.4.0  # Для моделирования транзитов
# pytransit>=2.5.0       # Для анализа транзитов
# emcee>=3.1.0           # Для MCMC анализа
# corner>=2.2.0          # Для визуализации MCMC